requirements.txt
bcrypt==3.2.2
beautifulsoup4==4.11.1
bs4==0.0.1
cachelib==0.7.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
Flask-Caching==1.11.1
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
pyOpenSSL==22.0.0
requests==2.27.1
soupsieve==2.3.2.post1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
